+++ ui/components/builder-header/RELEASENOTES.md
@@ -4,6 +4,14 @@
 
 <!-- ## [Unreleased] -->
 
+## 2.9.0
+
+### Fixed
+
+- Fixed issue where page type content was not being truncated
+  - Added `slds-has-flexi-truncate` to `slds-builder-header__item` element
+  - Added `<span class="slds-truncate" title="...">` inside of `<h1 class="slds-builder-header__item-label">`
+
 ## 2.7.0
 
 ### Added